refactor : vo 패키지 변경 (#809)
