+++ src/main/java/com/clubber/ClubberServer/domain/admin/implement/AdminTokenAppender.java
@@ -3,7 +3,7 @@ package com.clubber.ClubberServer.domain.admin.implement;
 import com.clubber.ClubberServer.domain.admin.domain.Admin;
 import com.clubber.ClubberServer.domain.admin.domain.AdminRefreshToken;
 import com.clubber.ClubberServer.domain.admin.repository.AdminRefreshTokenRepository;
-import com.clubber.ClubberServer.domain.auth.vo.TokenVO;
+import com.clubber.ClubberServer.global.jwt.vo.TokenVO;
 import com.clubber.ClubberServer.global.jwt.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
