+++ src/main/java/com/clubber/ClubberServer/domain/admin/service/AdminAuthService.java
@@ -7,7 +7,7 @@ import com.clubber.ClubberServer.domain.admin.implement.AdminReader;
 import com.clubber.ClubberServer.domain.admin.implement.AdminTokenAppender;
 import com.clubber.ClubberServer.domain.admin.implement.AdminTokenReader;
 import com.clubber.ClubberServer.domain.admin.validator.AdminValidator;
-import com.clubber.ClubberServer.domain.auth.vo.TokenVO;
+import com.clubber.ClubberServer.global.jwt.vo.TokenVO;
 import com.clubber.ClubberServer.global.config.security.SecurityUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
