+++ src/main/java/com/clubber/ClubberServer/domain/auth/implement/UserTokenAppender.java
@@ -2,7 +2,7 @@ package com.clubber.ClubberServer.domain.auth.implement;
 
 import com.clubber.ClubberServer.domain.auth.domain.UserRefreshToken;
 import com.clubber.ClubberServer.domain.auth.repository.UserRefreshTokenRepository;
-import com.clubber.ClubberServer.domain.auth.vo.TokenVO;
+import com.clubber.ClubberServer.global.jwt.vo.TokenVO;
 import com.clubber.ClubberServer.domain.user.domain.User;
 import com.clubber.ClubberServer.global.jwt.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
